Introducing abstract class
